refactor: 패키지 구조 변경 및 공통Exception 생성
